uae-legal-graphrag-neo4j-streamlit/requirements.txt
neo4j>=5.20
streamlit>=1.36
pydantic>=2
python-dotenv>=1
httpx>=0.27
tenacity>=8
pandas>=2.0
pytest>=7.0
